+++ clients/spot-quarkus/src/main/java/com/binance/connector/quarkus/spot/model/AllOrderListsRequest.java
@@ -51,7 +51,7 @@ public class AllOrderListsRequest extends BaseDTO {
 
     @JsonProperty(SERIALIZED_NAME_RECV_WINDOW)
     
-    private Double recvWindow;
+    private Long recvWindow;
 
     public AllOrderListsRequest() {}
 
@@ -131,7 +131,7 @@ public class AllOrderListsRequest extends BaseDTO {
         this.limit = limit;
     }
 
-    public AllOrderListsRequest recvWindow( Double recvWindow) {
+    public AllOrderListsRequest recvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
         return this;
     }
@@ -141,11 +141,11 @@ public class AllOrderListsRequest extends BaseDTO {
      *
      * @return recvWindow
      */
-    public Double getRecvWindow() {
+    public Long getRecvWindow() {
         return recvWindow;
     }
 
-    public void setRecvWindow( Double recvWindow) {
+    public void setRecvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
     }
 
@@ -194,7 +194,7 @@ public class AllOrderListsRequest extends BaseDTO {
             String limitValueAsString = limitValue.toString();
             valMap.put("limit", limitValueAsString);
         }
-        Double recvWindowValue = getRecvWindow();
+        Long recvWindowValue = getRecvWindow();
         if (recvWindowValue != null) {
             String recvWindowValueAsString =
                     DecimalFormatter.getFormatter().format(recvWindowValue);
