+++ clients/spot-quarkus/src/main/java/com/binance/connector/quarkus/spot/model/MyTradesRequest.java
@@ -64,7 +64,7 @@ public class MyTradesRequest extends BaseDTO {
 
     @JsonProperty(SERIALIZED_NAME_RECV_WINDOW)
     
-    private Double recvWindow;
+    private Long recvWindow;
 
     public MyTradesRequest() {}
 
@@ -181,7 +181,7 @@ public class MyTradesRequest extends BaseDTO {
         this.limit = limit;
     }
 
-    public MyTradesRequest recvWindow( Double recvWindow) {
+    public MyTradesRequest recvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
         return this;
     }
@@ -191,11 +191,11 @@ public class MyTradesRequest extends BaseDTO {
      *
      * @return recvWindow
      */
-    public Double getRecvWindow() {
+    public Long getRecvWindow() {
         return recvWindow;
     }
 
-    public void setRecvWindow( Double recvWindow) {
+    public void setRecvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
     }
 
@@ -271,7 +271,7 @@ public class MyTradesRequest extends BaseDTO {
             String limitValueAsString = limitValue.toString();
             valMap.put("limit", limitValueAsString);
         }
-        Double recvWindowValue = getRecvWindow();
+        Long recvWindowValue = getRecvWindow();
         if (recvWindowValue != null) {
             String recvWindowValueAsString =
                     DecimalFormatter.getFormatter().format(recvWindowValue);
