+++ clients/spot-quarkus/src/main/java/com/binance/connector/quarkus/spot/model/OpenOrderListsStatusRequest.java
@@ -30,11 +30,11 @@ public class OpenOrderListsStatusRequest extends BaseDTO {
 
     @JsonProperty(SERIALIZED_NAME_RECV_WINDOW)
     
-    private Double recvWindow;
+    private Long recvWindow;
 
     public OpenOrderListsStatusRequest() {}
 
-    public OpenOrderListsStatusRequest recvWindow( Double recvWindow) {
+    public OpenOrderListsStatusRequest recvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
         return this;
     }
@@ -46,11 +46,11 @@ public class OpenOrderListsStatusRequest extends BaseDTO {
      */
     
 
-    public Double getRecvWindow() {
+    public Long getRecvWindow() {
         return recvWindow;
     }
 
-    public void setRecvWindow( Double recvWindow) {
+    public void setRecvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
     }
 
@@ -84,7 +84,7 @@ public class OpenOrderListsStatusRequest extends BaseDTO {
         StringBuilder sb = new StringBuilder();
         Map<String, String> valMap = new TreeMap<String, String>();
         valMap.put("apiKey", getApiKey());
-        Double recvWindowValue = getRecvWindow();
+        Long recvWindowValue = getRecvWindow();
         if (recvWindowValue != null) {
             String recvWindowValueAsString =
                     DecimalFormatter.getFormatter().format(recvWindowValue);
