+++ clients/spot-quarkus/src/main/java/com/binance/connector/quarkus/spot/model/OrderAmendmentsRequest.java
@@ -52,7 +52,7 @@ public class OrderAmendmentsRequest extends BaseDTO {
 
     @JsonProperty(SERIALIZED_NAME_RECV_WINDOW)
     
-    private Double recvWindow;
+    private Long recvWindow;
 
     public OrderAmendmentsRequest() {}
 
@@ -135,7 +135,7 @@ public class OrderAmendmentsRequest extends BaseDTO {
         this.limit = limit;
     }
 
-    public OrderAmendmentsRequest recvWindow( Double recvWindow) {
+    public OrderAmendmentsRequest recvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
         return this;
     }
@@ -147,11 +147,11 @@ public class OrderAmendmentsRequest extends BaseDTO {
      */
     
 
-    public Double getRecvWindow() {
+    public Long getRecvWindow() {
         return recvWindow;
     }
 
-    public void setRecvWindow( Double recvWindow) {
+    public void setRecvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
     }
 
@@ -213,7 +213,7 @@ public class OrderAmendmentsRequest extends BaseDTO {
             String limitValueAsString = limitValue.toString();
             valMap.put("limit", limitValueAsString);
         }
-        Double recvWindowValue = getRecvWindow();
+        Long recvWindowValue = getRecvWindow();
         if (recvWindowValue != null) {
             String recvWindowValueAsString =
                     DecimalFormatter.getFormatter().format(recvWindowValue);
