+++ clients/spot-quarkus/src/main/java/com/binance/connector/quarkus/spot/model/OrderListPlaceOcoRequest.java
@@ -210,7 +210,7 @@ public class OrderListPlaceOcoRequest extends BaseDTO {
 
     @JsonProperty(SERIALIZED_NAME_RECV_WINDOW)
     
-    private Double recvWindow;
+    private Long recvWindow;
 
     public OrderListPlaceOcoRequest() {}
 
@@ -836,7 +836,7 @@ public class OrderListPlaceOcoRequest extends BaseDTO {
         this.selfTradePreventionMode = selfTradePreventionMode;
     }
 
-    public OrderListPlaceOcoRequest recvWindow( Double recvWindow) {
+    public OrderListPlaceOcoRequest recvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
         return this;
     }
@@ -848,11 +848,11 @@ public class OrderListPlaceOcoRequest extends BaseDTO {
      */
     
     
-    public Double getRecvWindow() {
+    public Long getRecvWindow() {
         return recvWindow;
     }
 
-    public void setRecvWindow( Double recvWindow) {
+    public void setRecvWindow( Long recvWindow) {
         this.recvWindow = recvWindow;
     }
 
@@ -1164,7 +1164,7 @@ public class OrderListPlaceOcoRequest extends BaseDTO {
             String selfTradePreventionModeValueAsString = selfTradePreventionModeValue.toString();
             valMap.put("selfTradePreventionMode", selfTradePreventionModeValueAsString);
         }
-        Double recvWindowValue = getRecvWindow();
+        Long recvWindowValue = getRecvWindow();
         if (recvWindowValue != null) {
             String recvWindowValueAsString =
                     DecimalFormatter.getFormatter().format(recvWindowValue);
